Renaming & some data type changes
- Get servers now returns IEnumerable and caches with a HashSet
- Devices online property is now a unsigned integer (was a signed long)

+++ LightVPN.Auth/Models/Server.cs
@@ -21,7 +21,7 @@ namespace LightVPN.Auth.Models
         public string CountryName { get; set; }
 
         [JsonPropertyName("devicesOnline")]
-        public long DevicesOnline { get; set; }
+        public uint DevicesOnline { get; set; }
 
         [JsonPropertyName("pritunlName")]
         public string FileName { get; set; }
